Fix length-prefix truncation divergence and RanSha reconstruction threshold
Validate declared payload lengths in u64 space before narrowing to
usize, so a crafted length can't diverge between 32-bit and 64-bit
targets. Also require shares from all n parties, not just 2t+1, before
RanSha reconstructs secrets.

+++ mpc/src/avss_mpc/input/input.rs
@@ -418,9 +418,9 @@ impl<F: FftField, R: RBC<Id = AvssSessionId>, G: CurveGroup<ScalarField = F>>
                 "Payload too short".to_string(),
             ));
         }
-        let declared_len = u64::from_le_bytes(msg.payload[..8].try_into().unwrap()) as usize;
+        let declared_len = u64::from_le_bytes(msg.payload[..8].try_into().unwrap());
         let input_len = self.client_data.lock().await.inputs.len();
-        if declared_len != input_len {
+        if declared_len != input_len as u64 {
             return Err(AvssInputError::InvalidInput(
                 "Mismatch in input and share length".to_string(),
             ));
