+++ mpc/src/avss_mpc/output/output.rs
@@ -107,8 +107,8 @@ impl<F: FftField, G: CurveGroup<ScalarField = F>> AvssOutputClient<F, G> {
                 "Payload too short".to_string(),
             ));
         }
-        let declared_len = u64::from_le_bytes(msg.payload[..8].try_into().unwrap()) as usize;
-        if declared_len != self.input_len {
+        let declared_len = u64::from_le_bytes(msg.payload[..8].try_into().unwrap());
+        if declared_len != self.input_len as u64 {
             return Err(AvssOutputError::InvalidInput(
                 "Declared input length does not match expected".to_string(),
             ));
